query/requirements.txt
qdrant-client==1.14.2
numpy>=2.1.0
redis==6.2.0
python-dotenv==1.1.0
beautifulsoup4==4.13.4
pandas>=2.1.0
nltk==3.9.1
tqdm==4.67.1
requests==2.32.4
httpx==0.28.1
grpcio==1.73.0
protobuf==6.31.1
psycopg2-binary==2.9.10
joblib==1.5.1
portalocker==2.10.1
click==8.2.1
certifi==2025.4.26